% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovr_ytd_compare.R
\name{ovr_ytd_compare}
\alias{ovr_ytd_compare}
\title{Calculates the year to date number against previous year to date}
\usage{
ovr_ytd_compare(
  dat,
  date_col_name,
  value_col_name,
  period,
  output_col_name = "compare",
  group_cols = NULL,
  skip_period = F,
  order_data = T,
  period_diff = 12,
  percentage = F,
  alt_year_col_name = NULL,
  output_cumsum = F
)
}
\arguments{
\item{dat}{A data frame or data.table}

\item{date_col_name}{The column name of the date column as a character string}

\item{value_col_name}{The column of the values to be evaluated as a character string}

\item{period}{Periodicity the data. Can be "month" or "year"}

\item{output_col_name}{The name of the column in the output data as a character string}

\item{group_cols}{Columns to group by when evaluating. Up to two are allowed}

\item{skip_period}{Whether to skip periods in the and go against the most recent period or stick rigidly to the previous period. If false value will be NA for previous period. Only applicable when using group columns. Can produce strange results if true.}

\item{order_data}{Whether the data should come out ordered or not. Can produce strange results if false.}

\item{period_diff}{Difference of periods to compare. For example, using "month" period, use 1 for month over previous month. 12 for month over previous year}

\item{percentage}{Whether the period comparison comes out as a percentage or a absolute value. Defaults to absolute value.}

\item{alt_year_col_name}{If using a different year to the calendar year, then the name of the year column to use instead. As a character string.}

\item{output_cumsum}{Output the cumulative sum column used for calculation. Used for tractability of calculation}
}
\value{
returns data frame or data table depending on the input data
}
\description{
Calculates the year to date number against previous year to date
}
\examples{


}
